emotionAnalysisFromText/requirements.txt
torch==2.6.0
numpy==2.2.2
transformers==4.50.0
dataset==1.6.2
evaluate==0.4.3
Flask==3.1.0